requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.0
flask-swagger-ui==4.11.1
future==0.18.3
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
python-whois==0.8.0
Werkzeug==3.0.1
